delete old manual and script plugins
